+++ pkg/server/middleware.go
@@ -1,0 +1,15 @@
+package server
+
+import "net/http"
+
+// RecoveryMiddleware is a middleware that recovers from panics and returns a 500 status code.
+func RecoveryMiddleware(next http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		defer func() {
+			if rec := recover(); rec != nil {
+				http.Error(w, "Internal Server Error", http.StatusInternalServerError)
+			}
+		}()
+		next.ServeHTTP(w, r)
+	})
+}
